+++ src/lib/utils/formatter/token.ts
@@ -14,6 +14,8 @@ export const formatDemimal =
     delimiter: boolean;
   }) =>
   (n: BigSource, fallbackValue: string): string => {
+    // TODO: fix after having unit test for this fn
+    // const num = big(n).toFixed(decimalPoints, big.roundHalfUp);
     const num = big(
       big(n)
         .mul(10 ** decimalPoints)